Added TypeScript support for $.fn
Added TypeScript support for $.fn

`[key: string]: Function;` is added to `Cash` such that #380 can be resolved.

`ArrayLike<HTMLElement>` is also added such that Cash can be array-like with HTMLElement results from `querySelectorAll`.

The element array shall be the normal use for this script. If the array content is other type such as document, window, number, etc, for typescript users, they shall do `$(document)[0] as any as Document`.